:sparkles: Add facts modal/code viewer & detail drawer tab  (#1057)
- Adds initial implementation of the facts modal, analysis detail drawer
tab for facts, and code viewer for facts json.

Notes: This does not cover the filter/sort logic for facts based on
sources. There is no mock data available from the api for this yet, but
we will be required to collate the fact sources ourselves on the UI side
based on the fact keys returned from the hub.
https://github.com/konveyor/tackle2-hub/pull/396 for more info on how
the api will look.

TODO: https://github.com/konveyor/tackle2-ui/issues/1061
TODO: [How should we type facts with an unknown data type?
](https://github.com/konveyor/tackle2-ui/issues/1062)

---------

Signed-off-by: ibolton336 <ibolton@redhat.com>