Add `SimpleDocumentViewerModal` to display analysis details
Instead of opening a new browser window to show a user a json
formatted analysis details document, open a modal on page with
an embedded and enhanced CodeEditor.

The `SimpleDocumentViewerModal` component allows fetching and
displaying the document as either a json or yaml document.  Users
can request the other format on screen.  By using the
`@patternfly/react-code-editor` component, a number of good things
are enabled:
  - the content is displayed on page and is code colored per language
  - the contents can be copied to the clipboard
  - the contents can be save to a file
  - the user can easily view either the json or yaml format of the
    same document, as provided by the rest api endpoint

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>