Refactor the client `tsconfig.json`
By updating the `tsconfig.json` to use `module: es2020`, this allows
webpack to do actual code splitting based on `React.lazy()`.

Many of the previous settings that were removed are now redundant
with either the `strict: true` or with the current `target`, `module`,
and `moduleResolution` settings.

The `ts-node` section has been added to allow the webpack configuration
files to be written in typescript.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>