:bug: Handle uploaded yaml with a single parsed value (#1081)
- Fixes a bug with the rules selector that was writing ruleset data to
each rule in the map. This caused all uploaded rules to show the ruleset
name rather than the rule name.
- Fixes an issue with yaml parsing. yaml.load returns an object when
only one value is found rather than an array when multiple values are
found from an uploaded rule file. This caused the app to blow up &
freeze when trying to iterate through the single object.
- TODO: Fix the enabled save button on edit with no dirty fields for the
custom-target-form #1073

Signed-off-by: ibolton336 <ibolton@redhat.com>