:bug: [main] Fix state management and validation of date pickers in migration wave modal (#1111)
Ports #1110 forward to main.

I implemented this fix first on the release-0.2 branch in anticipation
that it would be different for the PF4 and PF5 DatePicker components,
since it is a high priority for the release. However, this is a pretty
straightforward cherry-pick (with conflicts resolved) because the fix
required turned out to be the same for PF5.

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>