:sparkles: Issues table: Add description column, render expanded description as markdown, render links under markdown in table and incident viewer, add alert text to all incidents tab (#1116)
This PR contains four presentation tweaks to the Issues page from recent
discussions:
1. From a recent email discussion, it was decided that the although the
issue description can be long, the first line of it is the most
meaningful "title" for an issue or issue-based report and it should be
included in the table. This change adds the first line of the
description to the table, truncated with a tooltip.
2. The issue description in the expanded row content was being rendered
in plain text, and now its value can contain markdown. This change wraps
it using ReactMarkdown like the incident message.
3. The issue object (and now the issue-related reports as of
https://github.com/konveyor/tackle2-hub/pull/434) contains a `links`
array that we were not rendering. This change adds external links for
these URLs under the issue description in the expanded rows of the
issues table, and under the incident message in the incident viewer
modal.
4. There was a placeholder TODO string in the info alert above the "all
incidents" table in the incident viewer modal. This change adds a first
pass of that message: "Full details are only available for the first 5
incidents per file due to performance constraints". We can revisit this
if there are any objections to this text.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>
Co-authored-by: Ian Bolton <ibolton@redhat.com>