:bug: Fix aggregated wave status. Reflect all states in Jira instance  (#1124)
Resolves https://issues.redhat.com/browse/MTA-877
Resolves https://github.com/konveyor/tackle2-ui/issues/974
https://github.com/konveyor/tackle2-ui/issues/911

- Adds Completed status
- Adds In Progress status
- Filters out empty status from derived status calculation. This could
mean the issue has been deleted from the underlying JIRA instance OR
that the Jira ticket is in the process of being created. We have no way
of knowing how to interpret empty string, so this filters it out which
gives us a more accurate representation of aggregated status.
- Add TODO for info text around auth types for JIRA credential creation
- Remove Created By field in identity page table
- Fixes stale tickets data in migration waves table by mounting the
tickets query

Signed-off-by: ibolton336 <ibolton@redhat.com>