:sparkles: Remove empty name column for issues in favor of description column (#1152)
In @pranavgaikwad's demo we saw that the issue `name` field is not
useful and perhaps always empty in real data, and instead the first line
of the `description` field is the more meaningful identifier for a row
in this table. This change removes the name column and renames the
description column header to "Issue" as discussed.

![Screenshot 2023-07-17 at 1 48 36
PM](https://github.com/konveyor/tackle2-ui/assets/811963/98810f14-281d-4967-af33-bbc7b3f1629d)

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>