:bug: (various): styling issues with pf-v5-m vs pf-m (#1177)
## Missing Pencil on tables

closes #1166 

Also found several `pf-v5-m-` where the version should not be on the
modifiers. This affected many items, such as not adding the blue left
border to active menu items on the nav panel

Aside from removing the `-v5` from `pf-m` modifiers, for the table I
also removed the `inline-edit` div as it doesn't appear we are doing
inline edits and the styling for the inline-edit was basically just
being removed manually by the `enable` modifier.

should also closes #1165 

<!--
## PR Title Prefix

Every **PR Title** should be prefixed with :text: to indicate its type.

- Breaking change: :warning: (`:warning:`)
- Non-breaking feature: :sparkles: (`:sparkles:`)
- Patch fix: :bug: (`:bug:`)
- Docs: :book: (`:book:`)
- Infra/Tests/Other: :seedling: (`:seedling:`)
- No release note: :ghost: (`:ghost:`)

For example, a pull request containing breaking changes might look like
`:warning: My pull request contains breaking changes`.

Since GitHub supports emoji aliases (ie. `:ghost:`), there is no need to
include
the emoji directly in the PR title -- **please use the alias**. It used
to be
the case that projects using emojis for PR typing had to include the
emoji
directly because GitHub didn't render the alias. Given that `:warning:`
is
easy enough to read as text, easy to parse in release tooling, and
rendered in
GitHub well, we prefer to standardize on the alias.

For more information, please see the Konveyor
[Versioning
Doc](https://github.com/konveyor/release-tools/blob/main/VERSIONING.md).
-->

---------

Signed-off-by: gitdallas <dallas.nicol@gmail.com>