:bug: Fix PageSidebar structure and spacing of perpective selector above nav (#1185)
Fixes #1163

As part of the PF5 upgrade we failed to wrap the children of
`<PageSidebar>` in `<PageSidebarBody>` as mentioned [in the release
notes](https://staging.patternfly.org/get-started/upgrade/release-notes/):

> PageSidebar API has been updated. The nav property has been renamed to
children. Any content passed to the property should be wrapped in
PageSidebarBody.

Also, the padding at the top of the page sidebar body element has been
moved to the top of the nav element in PF5, so our custom styles for the
perspective selector broke. This fixes that CSS to restore the styles
established in https://github.com/konveyor/tackle2-ui/pull/251 (based on
[design guidelines
here](https://www.patternfly.org/v4/components/context-selector/design-guidelines#recommended-placement-0)).

Note that we may still want to follow up and check if these design
guidelines are still the best practice, since the context selector
component is deprecated and I can't find equivalent guidelines in the
new menu docs for a menu at the top of the nav like ours. It would be
nice to find a solution that is supported by the PF components without
these custom styles.

It's the little things 😄 ✨

![Screenshot 2023-07-21 at 12 07 14
PM](https://github.com/konveyor/tackle2-ui/assets/811963/2f0e852b-dd41-45cf-9143-8b115862589f)

Note: the diff is best viewed with "hide whitespace" checked in the diff
viewer settings. The only change to the JSX was nesting things under
`<PageSidebarBody>`.

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>