:bug: [backport release-0.2] Adjust Tags tab on Application Details Drawer when a Task is running (#1192)
Backport of #1191

When an analysis task is running, and the application drawer is open,
show both loading text and spinner and the existing tags. Once the
analysis is complete, the loading text/spinner will be removed and any
new tags will be shown.

Bug-Url: https://issues.redhat.com/browse/MTA-465

The Reports tab is also mentioned on MTA-465. This was already addressed
with #1103 (backported with #1190)

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>