:bug: Use truncated first line of issue description everywhere we were using issue name (#1186)
Followup to #1116

When we replaced the issue name property (possibly empty in real data)
with the truncated issue description in the table, we failed to replace
it everywhere else it is used.

This change uses the first line of the issue/report description where
necessary, and truncates it with a tooltip in places with limited
available width in case it is a long line.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>