:sparkles: Refactor useFetchTasks()'s select function (#1216)
The motivation for this refactoring came while working on #1191 and
trying to understand what tasks were given to the analysis table. The
select function used by `useFetchTasks()` to filter by addon and to keep
only the newest tasks per application id was confusing to read as
written. This refactor simplifies the steps involved to use a typical
stream processing approach.

Output of the refactored function is functionally identical to the
original implementation.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>