:seedling: Refactor useActiveRowState to allow number or string ids in state for consistency, misc cleanup (#1229)
I made these small cleanup changes before starting on refactoring
useSelectionState to fit the model of the other table-control hooks, but
that grew in scope so I broke this part into its own PR and will
continue the rest after PTO.

This PR:
* Removes commented-out `hasClickableRows` lines because I confirmed
that they are indeed no longer needed with PF v5
* Changes the behavior of `useActiveRowState` and
`getActiveRowDerivedState` so that the id of the active row as returned
in the `IActiveRowState` object can be a string or a number. This aligns
with other state hooks that deal with item ids, which can be strings or
numbers in the data model (as permitted by the type of `idProperty`
being `KeyWithValueType<string | number>`.
* Previously, when we added support for numeric ids in
https://github.com/konveyor/tackle2-ui/pull/893/files#diff-82ce09cd10287f153f994982c4f69732afc2b20f0a9a9426680bf5358101d146L6
this broke the behavior of `useActiveRowState` which was expecting only
string ids to be persisted. We worked around that by casting the id to a
string whenever it was passed to the hook's methods in
https://github.com/konveyor/tackle2-ui/pull/904/files#diff-82ce09cd10287f153f994982c4f69732afc2b20f0a9a9426680bf5358101d146L20,
but that was messy because it meant there were two layers of
serialization of these values: once when passed into the helper, and
again when serializing for `useUrlParams`.
* This change paves the way for the upcoming changes to
`useSelectionState` which will also allow both numbers and strings for
the ids of selected items in state.