:ghost: Target api updates (#1239)
UI changes relating to https://github.com/konveyor/tackle2-hub/pull/464

- Rename setting for ruleset/target order to match api changes
- Change TS model for Ruleset to match api changes. Ruleset is now a
field on the target model and can be appended to directly / read from
directly when performing CRUD operations against a custom target.
- Drive target card selection from label rather than ruleset/target
since we no longer need to store the entire ruleset/target state in the
wizard. This is because the analyzer no longer requires it to be passed
in the TaskData. We are now only passing labels. This simplifies the
wizard state & allows tracking of individual ruleset selection/ source
selection/ target selection to be removed.
- Now we can drive the select targets / select sources dropdowns on the
set-options analysis wizard form from labels rather than tracking
specific state for either of those.
- Changes format of default sources/targets to match new hub api label
format

---------

Signed-off-by: ibolton336 <ibolton@redhat.com>