:sparkles: Add component for downloading the static report (#1312)
- Adds a download button to handle downloading the analysis report with
custom headers driven by props passed in
- Uses the Fetch API to allow setting custom headers. Converts the
received data to a blob & then creates an object URL from that blob.
Then we can use that URL with a temporary anchor `<a>` element to
trigger the download with a localized loading state.
- This approach allows us to bin any custom headers we were adding in
the proxy and just use the fetch api to set the custom headers.

![image](https://github.com/konveyor/tackle2-ui/assets/11218376/3b30ee73-d5bc-4a1e-be32-4ca079321826)

Signed-off-by: ibolton336 <ibolton@redhat.com>