:sparkles: Remove pathfinder and convert assessment wizard to use new api  (#1315)
https://github.com/konveyor/tackle2-ui/assets/11218376/a8e7162b-24e3-4262-b04d-9969041e3976

This pull request aims to start the development of the single-app
application assessment flow. It includes the addition of a full flow for
the assessment process against the new data model, updates to match the
latest Go structs from the hub PR, removal of pathfinder-specific
elements, creation of msw mocks to handle assessment initial post and
patch operations, and a few other things.

Changes Made:

- Single-App Application Assessment Flow: Added a complete flow for
single-app application assessment. This flow includes handling initial
posts and patch operations.
- Model Updates: Updated the models to align with the latest Go structs
from the hub PR.
- MSW Mocks Creation: Created msw mocks to handle assessment initial
post and patch requests.

TODO:

- Comments Feature: Comments feature is currently commented out in the
assessments flow.
- Stakeholders & Stakeholder Groups: Identified that the code related to
stakeholders and stakeholder groups is not yet added to the assessment
struct on the API. Collaboration with the hub team is needed to
incorporate this functionality, as assessments can have associated
stakeholders.
- Mock Endpoint Enhancement: Need to add support for
/applications/:id/assessments when available
- Tests Update: Need to evaluate existing tests & update to match the
new data format for assessments.
- Assessment Risks Component: Will need to wire this up depending on the
future availability of the calculatedRisk assessment field from the API
- Hub Derived Fields for Assessments: Hub-derived fields for calculated
risk not yet available. These fields are planned to be calculated by the
hub after the final assessment submission, facilitating categorization
in the reports screen.
- Old Report Cards Consideration: Although the components for report
cards still exist in the repository, their usage in reports.tsx has been
commented out.

---------

Signed-off-by: ibolton336 <ibolton@redhat.com>