:sparkles: Implement delete archetype kebab action (#1333)
Part of #1264, followup to #1328.

![Screenshot 2023-09-06 at 4 43 16
PM](https://github.com/konveyor/tackle2-ui/assets/811963/f3dde2f4-22ff-4234-b591-dbfa8025cb8a)

Note that the change to the MSW mock is because that line was a typo
that resulted in a 404 when deleting an archetype, the delete
questionnaire endpoint is already mocked in its own file.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>