:sparkles: Show the name of objects to be deleted in the confirm delete modals (#1347)
Applies the pattern from
https://github.com/konveyor/tackle2-ui/pull/1333 to the rest of our
entity delete confirmation modals.

Note that in order to do this, I needed to refactor away an inconsistent
pattern: Some of our delete modals were kept track of with
`thingIdToDelete` as a number in state and `isConfirmDialogOpen` as a
boolean. In order to show the name in the confirm modal, this PR
refactors these incidences to store a reference to the whole object
`thingToDelete` instead of just the id. The `isConfirmDialogOpen` state
was also redundant because the confirm modal is always closed when
`thingToDelete` is falsy and open when it's truthy, so this PR also
removes those booleans from state and uses `thingToDelete` as the source
of truth.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>
Co-authored-by: Ian Bolton <ibolton@redhat.com>