Reorganize jest dependencies, no workspace `node_modules`
As a side-effect of having all of the jest dependencies only in
the `client/` workspace, a `client/node_modules` folder was being
created during npm install.  While this is ok for online upstream
builds, and downstream offline builds were breaking due to extra
layer of complexity.

Arranging dependencies across projects to prefer a single
`node_modules/` folder is highly preferred.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>